requirements.txt
aioca==1.7
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
bidict==0.23.1
bluesky==1.13.0a3
certifi==2024.2.2
charset-normalizer==3.3.2
cycler==0.12.1
dataclasses-json==0.6.4
dls-dodal==1.21.0
docopt==0.6.2
epicscorelibs==7.0.7.99.0.2
event-model==1.20.0
frozenlist==1.4.1
graypy==2.1.0
HeapDict==1.0.1
historydict==1.2.6
idna==3.7
importlib_resources==6.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nose2==0.14.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
ophyd==1.9.0
ophyd-async==0.3a1
p4p==4.1.12
packaging==24.0
pika==1.3.2
pillow==10.3.0
Pint==0.23
ply==3.11
pvxslibs==1.3.1
pydantic==1.10.15
pyepics==3.5.2
pyparsing==3.1.2
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
setuptools-dso==2.10
six==1.16.0
stomp-py==8.1.2
super-state-machine==2.0.2
toolz==0.12.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
websocket-client==1.7.0
workflows==2.26
yarl==1.9.4
zict==2.2.0
zocalo==0.32.0
